Merge in labctl and port controller logic
This is a completely new mechanism for loading environments and
equipment with some big and important differences from the other pycopia
backed mechanisms.

This means we need to port the existing controller logic to take
advantage of the newer APIs. It also means we can now merge some of the
controllers and equipment logic that was previously separated across the
two projects.