Support batched facts operations
Allow for calling `labctl facts --env <env name>` wherein the facts
for every location will be returned or can be modified (if you pass
key value pairs or --unset). This doesn't work with much else then the
table format at the moment but we should be able to fix it for YAML by
providing a custom formatter. We don't need it yet anyway.

Also, fail an assertion when we can't find the 'lab.yaml' config file in
the pwd.