Ensure we maintain reliable ssh during log capture
This adds a unit test which checks that even if ssh connections are
purposely closed during a test they are brought back up by the
logwatch plugin during the capture step at test teardown.

This required spinning up an alpine linux docker image running sshd to
simulate this in a local environment. Thus the new docker-py dependency.

QA-55 #comment This counts at the first test in the suite :)
QA-277 #In-Progress