Don't make --env required
Instead of always requiring the --env=NAME arg, set the 'anonymous'
env by default. Error with an informative message if the user tries
to look up an undefined role/equipment.

This allows a user to get useful output from `pytest -h` when run from the
root of the sng_qa repo without having to load an environment. Having
'anonymous' as the default env also turns out to be handy for testing.

QA-5 #In-Progress