Don't use a single connection across subclasses
The connection API using descriptors isn't compatible with our mixin
inheritance scheme. You have to declare a connection manually on each
controller class. We're not doing when relying on a mixin...

Instead keep a map of locations to proxies which has the added benefit
of caching connections per location.

QA-314 #In-Progress