Support pluggable environment providers
Add a new hook `pytest_lab_add_providers()` which is called from
`lab.provider.get_providers()` with a `ProviderManager`. This allows
hook implementations to register an environment provider and return
a config dict that it should be passed when called. Move all
provider and yaml config loading logic into `get_providers()` making it
the main entry point for the environment loading system.